requirements.txt
absl-py==0.11.0
aiohttp==3.6.3
aiosocksy==0.1.2
alembic==1.4.2
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bokeh==2.2.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cliff==3.5.0
cloudpickle==1.6.0
cmaes==0.7.0
cmd2==1.4.0
colorama==0.4.4
colorlog==4.6.2
cssselect==1.1.0
cycler==0.10.0
dataclasses==0.6
fake-useragent==0.1.11
fastapi==0.53.2
fastapi-login==1.5.1
fastapi-utils==0.2.1
finta==1.1
fsspec==0.8.4
future==0.18.2
google==3.0.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
grpcio-tools==1.33.2
h11==0.11.0
idna==2.10
inflect==5.0.2
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
lightgbm==3.0.0
liquidtap==1.0.1
lxml==4.6.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.2
more-itertools==8.6.0
multidict==4.7.6
multitasking==0.0.9
netifaces==0.10.9
numpy==1.19.1
oauthlib==3.1.0
optuna==2.3.0
packaging==20.7
pandas==1.1.1
parse==1.18.0
passlib==1.7.4
patsy==0.5.1
pbr==5.5.1
Pillow==8.0.1
pluggy==0.13.1
prettytable==0.7.2
protobuf==3.14.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.6.1
pyee==7.0.4
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
pyperclip==1.8.1
pyppeteer==0.2.2
pyquery==1.4.3
pytest==5.3.5
python-dateutil==2.8.1
python-editor==1.0.4
python-multipart==0.0.5
pytorch-forecasting==0.6.1
pytorch-lightning==1.0.8
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
simplejson==3.17.2
six==1.15.0
snownlp==0.12.3
soupsieve==2.0.1
sqlacodegen==2.3.0
SQLAlchemy==1.3.18
starlette==0.13.2
statsmodels==0.12.1
stevedore==3.3.0
supervisor==4.2.1
TA-Lib==0.4.19
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
torch==1.7.0
tornado==6.1
tqdm==4.53.0
tushare==1.2.62
twitter==1.18.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
uvicorn==0.12.3
w3lib==1.22.0
wcwidth==0.2.5
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
yarl==1.5.1
yfinance==0.1.55
sklearn==0.0
statsmodels==0.12.1
arch==4.15
bt==0.2.7
arrow~=0.17.0
typer~=0.3.2
uniswap-python==0.4.4
web3==5.15.0
